Added sending of HTTP/1.1 "Host:" header field to bitcoind
None of the bitcoind's except for bchd strictly enforce this, but RFC
2616 does specify that if this field is missing, the other end should
hang up. So, in the interests of correctness and having this code work
with future bitcoind's (as well as bchd should it implement persistent
connection), we added this field. It is now sent to bitcoind for
HTTP/1.1 POSTs (e.g. every single JSON-RPC request to bitcoind).