JSON over HTTP/1.1 (bitcoind client): Assert error if Connection: close
We need to more explicitly handle the situation with bchd where it is
doing Connection: close on us so that users that attempt to use Fulcrum
with bchd know why Fulcrum is not happy.

We added a more descriptive error message about why we reject the remote
bitcoind. This error message informs the user that Fulcrum requires
persistent HTTP/1.1 as described in RFC 2616.

The rationale for this is that the entire mechanism we devised for
talking to bitcoind in a performant manner would suffer if we had to
handle having to reconnect for each of the (potentially millions) of
requests we need to be able to send to bitcoind.

I'm working with Josh on updating bchd to support persistent
connections, so this error message will hopefully only occur for users
in the future who are running older versions of bchd.

Also in this commit: A potential bugfix if a header value containted ':'
characters (they were being transformed to ' ' spaces).