Bugfix: Qt was setting the locale, which broke JSON serialization
Oops!  I didn't realize the QCoreApplication object would mess with the
C locale, and our JSON serializer relies on the C lib to serialize
floats.  As such it broke JSON serialization for locales like
LANG=de_DE, etc.

This fixes that bug.  Thanks for Georg Englemann for reporting this bug.