Log filter: Use a better, more compact key for the filter
We were using the first 1024 bytes of the tx as the key -- in an
optimistic attempt to not hash anything on braodcast unless there's
a failure.

But since Mist Mining is so prevalant, a better approach is to sha256 all
the tx bytes and use the first 16 bytes of the resultant hash as a txkey
for the broadcast failure bloom filter.

This cuts processing time by up to 64x when we must consult the bloom
filter on broadcast fail (which is a common occurrence now).

tl;dr: the bloom filter really needs short keys.. it cannot deal well
with very long keys (performance hit is high since it must hash the data
20-50 times, depending on fpRate setting).