Simplified and reduced the Connection: close handler to just warn
It turns out aborting the app on Connection: close was a mistake.
bitcoind may send this header if it is shutting down, or it may send it
if it is warming up (BU does this).

Instead we simplified the handling and just print a warning message only
if statis == 200 and Connection: close.

Thanks to Axel Gembe for reporting this issue.