Handle bchd quirks so we can synch on bchd, plus fixed a subtle RPC bug
- bchd does not send the "initialblockdownload" key for `getblockchaininfo`
so we do NOT want to abort the app if this key is missing, but simply
tolerate the situation.

- The network names bchd uses "mainnet" and "testnet3" differ from what
we expect, so I added code to deal with that situation and map them to
the proper BTC::Net enum type.

- I found a bug in the RPC handling code where the idMethodMap method
wasn't being propagated properly to the message.method if the message
type was "error". This bug wasn't breaking any code -- just that future
code might wish to inspect this field to determine which method failed
with an error.