Redid the way bitcoind requests work -- made them more efficient
- Previously we were connecting the "gotMessage" signal from each
bticoind to all request objects -- which would mean basically that there
would be an O(N) scan through all active request contexts when a
response arrived from bitcoind.  We instead now use a dispatch table
that is keyed off request ID. So when messages arrive from bitcoind
there is an O(1) lookup in the table and the dispatch is done only to
the context object in question.  This should save CPU cycles .. especially
under load.

- Redid the object lifecycle a little bit for the ReqCtxObj objects --
only the "sender" object can keep them alive -- everything else keeps a
weak reference to them and if the "sender" object dies -- the context
object is deleted.

- Various bits of refactoring in BitcoinD.cpp -- overall the bitcoind
request scheme is much more efficient and cleaner now.