BitcoinDMgr: Handle request timeouts to bitcoind (rare but possible)
It turns out upon more careful inspection of the code that requests to
bitcoind have a kind of race condition if the bitcoind connection goes
down in the time that we grabbed the BitcoinD pointer but before we
could register a "lostConnection" signal handler for the request.

Also, it's possible for bitcoind to never reply (theoretically). We need
to handle these two cases and issue timeout responses to the sending
object if this occurs -- so that they at least get a reply and to
also guarantee the eventual deletion of the "ReqCtxObj" object.

The timeout is 15 seconds (in practice it may take over 20 seconds for
the timeout message to be delivered to the request sender).

The timeout mechanism is implemented by periodically checking the extant
request table and simply timing out old requests that are >15 seconds
old.