WebSocket: Correctly handle multiple values in the Connection header
Some user agents send a comma separated list of values in the
connection header so this needs to be handled correctly. It is
enough if one of the values is "upgrade" now.

This also moves the header checking code into the ClientServerBase
to consolidate the checks for client and server.