bitcoind RPC: Tolerate possible bitcoind RPC "Connection: keep-alive" header
Since 'keep-alive' is what we really semantically want anyway (and is the default if
missing), we should be tolerating this header field from bitcoind and not even warn or
print debug info if it occurs.

Currently no bitcoind that I know of sends this header field, but some proxies like
nginx *may* send it.  See issue #28.