Removed the periodic printing of the Fulcrum "banner" to the log
This was a holdover from the days when the log was all "uptime" based
and the banner was needed to synch the app "uptime" to wall-clock time.
Printing of this banner is not needed and if anything causes confusion
(I have had 1 user tell me they thought Fulcrum had restarted
spontenaously when seeing the banner appear).