Fixed Windows segfault-before-main if comiled with -DENABLE_TESTS
It turns out for some reason initialization order was wrong on Windows
and ths std::map instances for registeredTests and registeredBenches
were being accessed before their construction.

In order to fix this we are using std::unique_ptr for these types and
the first time through the common "register" function we simply
construct the maps via std::make_unique<>()..

This fixes the crash-on-startup bug seen on Windows when compiling with
-DENABLE_TESTS.

We also enabled the tests on travis for windows.