qmake: Added optional suppression of rocksdb auto-detection
This was needed because the redhat package that was building fulcrum was
accitendally saying "rocksdb: using static lib" when they really were
using the system lib.  To fix this we just offer the capability that we
disable auto-detection altogether for packages such as .rpm optionally,
if "LIBS" contains "-lrocksdb" when qmake is invoked.

So for example on Linux Fedora, from the redhat package build script:

$ qmake Fulcrum.pro "LIBS+=-lrocksdb -lz -lbz2"

The above will see that "-lrocksdb" is present already in the LIBS
variable when qmake was invoked, and would skip any rocksdb
auto-detection.  The assumption is the calling script set up everything
with the LIBS variable itself, so there is no need to do any
autodetection.

Optionally the calling script can also add INCLUDEPATH+= too to the args
in case rocksdb headers live in a non-standard place.