Added `simdjson` CLI arg and conf file variable
If enabled, and if the app supports it at runtime (and it it has
simdjson compiled-in), then simdjson will be used for all JSON-RPC
parsing.  It is an extremely fast backend (at *least* 2x faster and
sometimes even faster than that on large data sets).

Since it is a new backend and it hasn't been war-tested yet, we will
default it to off.

This commit also contains various small bits of code cleanup as well.