Tweaks to the thread used to detect signal handler-initiated exits
We use a QThread now as opposed to std::thread for the "exit thread".
 Rationale: QThread's API is richer and we can set the thread stack size.

Since this thread does almost nothing, we want to set it to have a very
minimal thread stack size so as to avoid wasting virtual address space
for a thread that probably will never use more then ~200 bytes of stack.

As such we added generic platform-neutral code in Util.cpp to detect the
minimum stack size on the system the code is running on, and to set the
ExitThr's stack size to this size.

ExitThr is now a nested private class, implemented inside
App::startup_Signalhandlers(). I love this language feature. :)