Fixup for PTHREAD_STACK_MIN not being accurate in static builds
This is because we build on 1 system but potentially run on others with
different libs.  PTHREAD_STACK_MIN may not be accurate in that case. So
we probe the minimum stack size by actually attempting to start a thread
with that stack size set, and if it succeeds, we can assume the minimum
given by PTHREAD_STACK_MIN is accurate.

If it does not succeed, we just can proceed with the default stack size,
as the "minimum" stack size, which should always work.