Made signhandler setup add the deferred restore for SIGHUP
Just for symmetry -- we made SIGHUP go from SIG_DFL -> SIG_IGN;
on app destruction we should restore it like we do the other signals,
that is, go back it previous action.

Also made the signal reg/unreg code store the previous signal's action
and restore THAT, rather than go with SIG_DFL. This is also a symmetry
thing and not strictly necessary in this app.