Refactor / nit for AbstractConnection::setMaxBuffer
We no longer clamp the passed-in value. Instead, client code must ensure
its sanity, which is the case in this codebase.  This allows us to have
the BitcoinD class instances pass very large max buffers directly in their
constructors (sizes that would normally be rejected by the claming code).

We got rid of that noClampMax arg which was inelegant.