Added CLI arg --bd-clients (conf file var: bitcoind_clients)
We no longer hard-code the number of bitcoind clients that we spawn to
connect to bitcoind. It was previously hard-coded as 3. Now, we allow
the user to configure this setting.

In my testing, on latest patches to BCHN, increasing this value above 3
(and also increasing -rpcthreads= in bitcoind) helps improve synch time.