Adjust some param defaults: db_max_open_files & bitcoind_timeout
We set the db_max_open_files to default 30 rather than 100. This should
be a decent default for low-end systems and performs ok on high end as
well.  The default of 100 could potentially eat RAM over long Fulcrum
runs and some users may not want that.

Also the default bitcoind_timeout has been bumped up to 30 seconds from
20 -- if bitcoind is timing out we should wait for it more.  In my testing
on an HDD with a low-end system on scalenet 20 seconds was a bit too
close for comfort but 30 seconds is decent and never hit.