Mempool performance: Don't hold the exclusive lock when reading from DB
Eureka! I figured it out. The mempool synch task was holding the
exclusive lock on the mempool while reading in utxo info from the db.
This turns out to be terribly slow when first building a huge mempool
(such as on BTC) and led to the app freezing for seconds at a time.

This is unacceptable!

Instead, we pre-cache the coins we will need with the mempool locked in
read-only mode. So we read from the DB while other threads are able to
read from the mempool just fine.

The actual update to the mempool then is done with a writer lock
(exclusive lock) on the mempool but the coins are read from memory cache
rather than (slowly) from the DB.

This should improve concurrency and performance on big slow mempools
such as on BTC.