Made the mempool bench test much more accurate
But also slowed it down significnatly. It now takes a deep copy of the
TxRefs and after the "drop any tx" test runs, a second mempool is built
only via adds.  The two mempools ae depp-compared for equality (which
runs very slow, but at least it tests thoroughly).

Also in this commit:

- Added the Mempool::deepCompareEqual() function (very slow) for tests.

- Added the "Tic" class to stop having to write timing/profiling
  boilerplate code.  It manages taking a timestamp.

- Made the getTime*() functions noexcept

- Made the AsyncSignalSafe::writeStdErr function noexcept