Implemented confirmedInBlock; plus other fixups and refactoring (#68)
* Implemented confirmedInBlock; plus other fixups and refactoring

This further optimizes processing for full mempools. In the addBlock()
execution path, we were removing too many tx's before.  Instead, now we
remove just the ones in the block and not their descendants.  For
descendants, we recategorize the spends as "confirmed spends" for txids
that are now in the new freshly-confirmed block.

* Tally sort time in nanos in Mempool::rmTxsInHashXTxs

* Fixed a comment

* Prepare for merge: put some debug output as TRACE, disable ENABLE_TESTS