Storage.cpp: remove potentially unaligned memory accesses
In particular in deserialization code, we were doing things like casting
char * buffers to a struct *.  While this works ok on Intel or modern
ARM, this is not portable.  Older ARM processors may have a bad time
with unaligned access.

In addition, this is also officially undefined behavior in C++.

Now in this commit, we instead copy all char * buffer data around to their
destinations using std::memcpy, casting through std::byte * to be safe with
C++ aliasing rules. In C++ this is the standard way to do such things.