Added conf file setting: max_reorg
This allows the user to configure how many blocks of undo data Fulcrum
will save to the db.  The default is 100 (previous versions of Fulcrum
had the max_reorg internally hard-coded as 100).  Users may set this to
any size from 100 to 500,000.

The default is fine -- setting it larger allows for larger reorgs to
occur without needing to resynch, at the expense of some disk space.

Code/logic was also added to deal with undo block "holes" which may now
arise in the unlikely even that the user configures max_reorg > 100 with
this version, then switches back to an older version of Fulcrum (which
still uses the 100 hard-coded undo depth size), then switches back to
this version.