Added conf option: `txhash_cache` (specifies max memory for lru txhash caches)
Previously the caches basically used a hard-coded size of 256MB (note we
calculated the size wrong -- we hard-coded 200MB but it actually was
using max 256MB).

This commit fixes the size calc to be more accurate (by taking into
account the extra QByteArray overhead we didn't take into account before),
and it also exposes a new knob to the user: txhash_caches.  User can now
control the size of the "LRU TxHash -> TxNum" and "LRU Height ->
TxHashes" caches.

Also in this commit: Make CostCache use a unique_ptr to be more
idiomatic.