Controller: Ensure bitcoind is using txindex, and if not, warn the user
We do this by asking bitcoind for tx #1 (the first tx we know about
after genesis).  If bitcoind replies with a valid tx, it has txindex
enabled.

If it does not -- that indicates to us that it is not using txindex, so
we warn the user about the situation and ask them to enable txindex.

Also in this commit: don't convert tx response from bitcoin daemon
into QString.  Leave it as QByteArray.  This is likely going to save
cycles and allocations (it already comes in off the wire as QByteArray
and is placed in our QVariant as QByteArray, so converting to QString
was wasted cycles).