Allow tor_proxy to be a hostname:port pair (in addition to IP:port)
Previous to this commit, the tor_proxy in config was interpreted as an
IP address string only.  Hostnames were not supported.

This commit adds support for looking up hostnames for tor_proxy (this is
useful in a docker setup, see #71).

If the user specifies a hostname for tor_proxy, it will first be
interpreted as an IP address, and if that fails, we fall back to (blocking)
hostname lookup.  This happens once during app init, and after the
hostname is resolved to an IP address, the IP address is what is stored
for tor_proxy in the global Options object.

This closes issue #71.