Fix checkdb to tolerate known inconsistencies on MainNet chain
The BCH/BTC mainnet chain has dupe tx's for the following txids:

- d5d27987d2a3dfc724e359870c6644b40e497bdc0589a033220fe15429d88599
  (appeariing as coinbase for 91842 and 91812)

- e3bf3d07d4b0375638d5f1db5255fe07ba2c4cb067cd81b84ee974b6585fb468
  (appearing as coinbase for 91880 and 91722)

This was throwing off our "checkdb" routine (-C --checkdb), causing it
to trigger a false positive on DB corruption.  There is no DB corruption
here! Bitcoin itself just has an inconsistency.

We created exeptions for those two outpoints and they are allowed to be
"off" with respect to the scripthash-unspent table.  (2 entries in
scripthash unspent but only 1 entry in utxoset!).

Before this commit --checkdb would always fail with a false positive.
Now it works as expected.