Mempool synch: Be more tolerant of dropped tx's as we synch
This fixes issues for in particular BTC where on first synch, we may
have to download 100k tx's when synching the mempool.

During this time (it can take 30-60 sec!) -- it's possible for tx's in
our list that we got from `getrawmempool false` to drop out of the
mempool -- and thus they return an error reply when asking for them via
`getrawtransaction`.

Previous to this commit we would panic and restart the synch task, which
was bad for BTC installs since it meant there was no guarantee the task
would ever successfully complete in pathological cases with high mempool
pressure!

Instead, it's actually better to just ignore the tx's we fail to
download since they are gone from mempool anyway, and continue synching
what we can.  A gone tx is as if it never existed, so it does no harm.
Aborting the task or restarting it as we did before was too cautious of
us.

After this change, BTC mempool synchs are much more reliable and less prone
to spurious excessive failures and restarts.

This should improve app performance overall on BTC, and even for BCH
should we ever get mempool pressure.

Also in this commit:

- cleaned up code for CtlTask::ts to use the new Tic class
- updated SynchMempoolTask to actually update its "lastProgress" stat
  for /stats endpoint (oops! forgot to update this stat in the initial
  implementation).
- SynchMempoolTask logs its total elapsed time on destruct if over 1.0
  secs for debug purposes.

Still left TODO in future: see about parallelizing the mempool tx download --
although the cost of mempool synch is dominated by db lookups and building
the Mempool data structure, it still should help to parallelize the downlaods.
(I haven't done this yet because it adds complexity and potential failure modes
to do this -- but it's a potential optimization we can do here).